study(step2): 카드 숫자, 문양에 대한 객체 추가
