+++ src/main/kotlin/blackjack/domain/CardNumber.kt
@@ -1,0 +1,19 @@
+package blackjack.domain
+
+enum class CardNumber(val rawValue: String) {
+
+    Ace("A"),
+    Two("2"),
+    Three("3"),
+    Four("4"),
+    Five("5"),
+    Six("6"),
+    Seven("7"),
+    Eight("8"),
+    Nine("9"),
+    TEN("10"),
+    King("K"),
+    Queen("Q"),
+    Jack("J")
+    ;
+}
