+++ src/main/kotlin/blackjack/domain/CardNumbers.kt
@@ -1,0 +1,6 @@
+package blackjack.domain
+
+class CardNumbers {
+
+    val cardNumberList: List<CardNumber> = CardNumber.values().toList()
+}
