+++ src/main/kotlin/blackjack/domain/Suit.kt
@@ -1,0 +1,10 @@
+package blackjack.domain
+
+enum class Suit(val rawValue: String) {
+
+    Spade("스페이드"),
+    Diamond("다이아몬드"),
+    Heart("하트"),
+    Clover("클로버")
+    ;
+}
