+++ src/main/kotlin/blackjack/domain/Suits.kt
@@ -1,0 +1,6 @@
+package blackjack.domain
+
+class Suits {
+
+    val suitList: List<Suit> = Suit.values().toList()
+}
